Move should_write_[electronic, positronic]_results_to_file to utils.py
